Add types for circular progress
